# Solver: clean up and optimise `SolveBySubstitution()` and `WriteJacobian()`