--- a/src/dsc.h
+++ b/src/dsc.h
@@ -36,6 +36,16 @@ operator<(T const &lhs, T const &rhs) {
     return lhs.v < rhs.v;
 }
 
+template<class T>
+struct HandleHasher {
+    static_assert(IsHandleOracle<T>::value, "Not a valid handle type");
+
+    inline size_t operator()(const T &h) const {
+        using Hasher = std::hash<decltype(T::v)>;
+        return Hasher{}(h.v);
+    }
+};
+
 class Vector;
 class Vector4;
 class Point2d;