This PR cleans up and optimises various parts of the solver, especially around `SolveBySubstitution()` and `WriteJacobian()`. The latter can result in noticeable speedups on some models (see initial numbers in my comment [here](https://github.com/solvespace/solvespace/pull/1563#issuecomment-2807422537)).

-----
 #### Original PR description (superseded)

~While trying to look into #1247 I was having trouble reproducing the issue, and the model was failing in weird ways (such as multiple constraints not satisfied, causing horizontal and vertical lines to become diagonal).~

~After staring hard at the code for a long time, I finally traced the issue down to commit 708a08f04b910337965859249b87684696369219 (merged in #1159) which made changes in order to speed up solving by substitution. However, the changes inadvertently  caused substitution to persist across solve runs and across groups, which broke multiple cases.~

~There are multiple options for fixing it, and I chose to do it by removing the `substd` member from `Param` (in order to prevent
such breakage in the future), and replacing the logic with similar runtime complexity logic that doesn't require modifying `Param` itself.~