@iscgar 2025-04-08
CI is failing due to spurious 502 errors when cloning the FreeType repository. The job should be retired, but I have no permission to do so. cc @ruevs.