@iscgar 2025-04-10
No, it does not fix #1247. I only encountered the problem this PR is intended to fix while trying to debug that other issue.

However, please hold off integrating this PR for now, because I'm currently unable to reproduce the issue. Also, in hindsight my description of it doesn't make much sense, becuase `System::SolveBySubstitution()` only works on copies of the parameters that are held by the system, so they should not persist across solve runs, and most definitely not across groups.

While I still think this PR brings some measurable improvement (such as reducing memory usage by getting rid of the `substd` member of `Param`, which is only ever used by substitution, and better algorithmic complexity for `Expr::UsedParams()` through the usage of `std::unordered_set`, as well as better algorithmic complexity by avoidance of traversing the substitution linked list), I want to try and reproduce the problem that I encountered in order to understand what caused it and ensure that it is actually fixed.