@iscgar 2025-04-15
I'm still unable to reproduce the issues I was having, so I'll have to chalk it up to a broken build on my machine.

This PR's description is now incorrect, as it turned into a bunch of small cleanups and optimisations, with a focus on `SolveBySubstitution` and `WriteJacobian`. I'll update the description to match the new state of things soon, but I'm marking it as ready to review in the mean time, so please take a look and let me know what you think.

Depending on the model used, different parts of the code are exercised, so the overall speedup will vary as well. The following list tallies the improvement in runtime (in release mode with LTO) for two operations: generating groups (with DoF calculation suppressed; depending on model this can exercise either of `SolveBySubstitution()` and `WriteJacobian()`, or both), and marking free points (which exercises `WriteJacobian()`):

* 8040_finished_constrained
  * Generate
    * Before: ALL(BB)=57ms, ALL=85ms, DIRTY=33ms
    * After: ALL(BB)=52ms, ALL=79ms, DIRTY=31ms
    * Change: insignificant
  * Mark free
    * Before: 927ms
    * After: 710ms
    * Change: -24%
* case-vents
  * Generate
    * Before: ALL(BB)=36ms, ALL=1247ms, DIRTY=1132ms
    * After: ALL(BB)=33ms, ALL=1243ms, DIRTY=1118ms
    * Change: insignificant
  * Mark free
    * Before: 1323ms
    * After: 1290ms
    * Change: insignificant
* Chapiteau vue de dessus
  * Generate
    * Before: ALL(BB)=120ms, ALL=129ms, DIRTY=not measured
    * After: ALL(BB)=118ms, ALL=126ms, DIRTY=not measured
    * Change: insignificant
  * Mark free
    * Before: 4985ms
    * After: 3970ms
    * Change: -20%
* constrainfreeze
  * Generate
    * Before: ALL(BB)=60ms, ALL=61ms, DIRTY=not measured
    * After: ALL(BB)=55ms, ALL=55ms, DIRTY=not measured
    * Change: insignificant
  * Mark free
    * Before: 245ms
    * After: 125ms
    * Change: -49%
* LeliaManya2
  * Generate
    * Before: ALL=not measured, DIRTY(BB)=493ms, DIRTY=836ms
    * After: ALL=not measured, DIRTY(BB)=491ms, DIRTY=830ms
    * Change: insignificant
  * Mark free
    * Before: 719ms
    * After: 706ms
    * Change: insignificant

All of my tests were done with DoF calculation suppression, and yet `SolveBySubstitution()` has essentially no impact, and the changes which were originally motivated by what I perceived as a bug are now merely a cleanup which may only benefit models with many substituable equations (I was having trouble coming up with such a model, so I'm not sure this is actually useful).