@phkahler 2025-04-16
@iscgar  These look pretty good. I'm not sure I like 
[this one](https://github.com/solvespace/solvespace/pull/1563/commits/a6a4da833572af071f7ad4df660a087b55ce2263) because it adds extra complexity and the solver is rarely a performance bottleneck. How has it been for you to learn the solver code and work with it? Would that change make such work harder?


Unrelated, there has been discussion in the past about solving more than one group at a time. Maybe by adding a checkbox to a group "solve with previous group" which could be chained to solve several together. This is not an existing feature, but I'm wondering how possible you think this might be, and if maybe the unused "tag" was related?