@iscgar 2025-04-16
> @iscgar These look pretty good. I'm not sure I like [this one](https://github.com/solvespace/solvespace/pull/1563/commits/a6a4da833572af071f7ad4df660a087b55ce2263) because it adds extra complexity and the solver is rarely a performance bottleneck. How has it been for you to learn the solver code and work with it? Would that change make such work harder?

Yeah, I can't say I like it either, but without it `SolveBySubstitution()` has `O(n^2)` complexity in case there are many independent substitutions, and I just couldn't let it stand considering the original Evil-Spirit commit message contained the words "linear complexity" :)

I find the solver code quite easy to work with, since it's mostly a direct implementation of the textbook Gauss-Newton algorithm. I initially started looking into `SolveBySubstitution()` because it was the only exception, and it took me a moment to figure out what it was doing with dragged parameters and cycle breaking. That specific change is still easier to understand IMO than the original implementation, as everything directly points to the target instead of traversing a linked list, but it is definitely more complex than my initial change in the commit preceding it (which is why I added a bit too many comments to explain why the code works).

As I was unable to come up with a model where `SolveBySubstitution()` would be the bottleneck in the solver, I'm fine with dropping that change if you feel that it isn't worth the added complexity.

> Unrelated, there has been discussion in the past about solving more than one group at a time. Maybe by adding a checkbox to a group "solve with previous group" which could be chained to solve several together. This is not an existing feature, but I'm wondering how possible you think this might be, and if maybe the unused "tag" was related?

Solving groups together should be easy enough to implement in the solver, as it's just a matter of using more than a single value in filtering constraints and entities to use for writing the equation system. However, from a UI standpoint it would probably be harder to reason about, with rank test suppression and e.g. bad constraints from multiple groups. I'm curious though: what's the advantage of solving multiple groups together? Is there a specific use case where it can potentially result in a noticeable speedup?

The `tag` is only an internal solver parameter for separating subsystems, so it's unlikely to have any connection with solving multiple groups, and is just a leftover from the pre-Eigen solver (originally the Jacobian was written inside `NewtonSolve()`, and this was changed in bab13b821f5fc874463eeb8d0b7c438100373aba).